Refactored code to be more inline with the ReadMe as well as imported styled-components
