+++ instagram/package.json
@@ -5,7 +5,8 @@
   "dependencies": {
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
-    "react-scripts": "2.1.3"
+    "react-scripts": "2.1.3",
+    "styled-components": "^4.1.3"
   },
   "scripts": {
     "start": "react-scripts start",
